almost finished with the analysis function

+++ cogs/spotify.py
@@ -11,6 +11,7 @@ import requests
 import os
 import matplotlib.pyplot as plt
 from palettable.colorbrewer.qualitative import Pastel1_7
+import datetime
 
 class Spotify(commands.Cog):
     
@@ -284,26 +285,36 @@ class Spotify(commands.Cog):
             await ctx.send(embed=self.create_connect_embed())
         else:
             track = sp.internal_call('/v1/me/player/currently-playing', access_token)
-            features = sp.internal_call(f'/v1/audio-features/{track["item"]["id"]}', access_token)
-            names = f'danceability ({features["danceability"]})', f'energy ({features["energy"]})', f'speechiness ({features["speechiness"]})', f'acousticness ({features["acousticness"]})', f'liveness ({features["liveness"]})', f'valence ({features["valence"]})'
-            size = [features['danceability'],
-                features['energy'], 
-                features['speechiness'], 
-                features['acousticness'], 
-                features['liveness'], 
-                features['valence']]
-            center_circle=plt.Circle( (0,0), 0.7, color='white')
-            plt.pie(size, labels=names, colors=Pastel1_7.hex_colors)
-            p=plt.gcf()
-            plt.gca().add_artist(center_circle)
-            plt.title(f"Spotify catalog information for: {track['item']['name']} by {track['item']['artists'][0]['name']}")
-            plt.savefig('chart.jpeg')
-            file = discord.File('chart.jpeg', filename='chart.jpeg')
-            content = discord.Embed(colour=int(util.color_from_image('https://www.scdn.co/i/_global/touch-icon-72.png'), 16))
-            content.set_image(url='attachment://chart.jpeg')
-            await ctx.send(file=file, embed=content)
-            os.remove('chart.jpeg')
-            plt.clf()
+            if not track:
+                await ctx.send("need to the bit in for when you're not listening to something")
+            else:
+                features = sp.internal_call(f'/v1/audio-features/{track["item"]["id"]}', access_token)
+                names = 'danceability', 'energy', 'speechiness', 'acousticness', 'liveness', 'valence'
+                size = [features['danceability'],
+                    features['energy'],
+                    features['speechiness'],
+                    features['acousticness'],
+                    features['liveness'],
+                    features['valence']]
+                center_circle=plt.Circle( (0,0), 0.7, color='white')
+                plt.pie(size, labels=names, colors=Pastel1_7.hex_colors)
+                p=plt.gcf()
+                plt.gca().add_artist(center_circle)
+                plt.title(f"Spotify catalog information for: {track['item']['name']} by {track['item']['artists'][0]['name']}")
+                plt.savefig('chart.jpeg')
+                file = discord.File('chart.jpeg', filename='chart.jpeg')
+                content = discord.Embed(colour=int(util.color_from_image('https://www.scdn.co/i/_global/touch-icon-72.png'), 16))
+                content.set_image(url='attachment://chart.jpeg')
+                content.set_author(name=f'📊{track["item"]["name"]} feature analysis',
+                    url=track['item']['external_urls']['spotify'],
+                    icon_url='https://www.scdn.co/i/_global/touch-icon-72.png')
+                content.add_field(name='Key:', value=f'`{util.get_pitch(features["key"])} {util.get_mode(features["mode"])}`')
+                content.add_field(name='Length:', value=f'`{util.stringfromtimestamp(features["duration_ms"]/1000.0)}`')
+                content.add_field(name='Tempo:', value=f'`{int(features["tempo"])} beats per minute`')
+                content.set_footer(text='🎵 Analysis provided by Spotify')
+                await ctx.send(file=file, embed=content)
+                os.remove('chart.jpeg')
+                plt.clf()
 
     # helper functions
